webapi/tct-filesystem-tizen-tests/filesystem/FileStream_write_exist.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Kaiyu <kaiyux.li@intel.com>
        Guan, JingX <jingx.guan@intel.com>

-->

<html>
<head>
<title>FileStream_write_exist</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/filesystem_common.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: FileStream_write_exist
//==== LABEL Check if FileStream::write() method exists
//==== PRIORITY P0
//==== SPEC Tizen Web API:IO:Filesystem:FileStream:write M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/filesystem.html
//==== TEST_CRITERIA ME

var t = async_test(document.title),
    fsTestFileName = getFileName("existFile08.txt"),
    resolveSuccess, resolveError, openStreamSuccess, openStreamError, file;

t.step(function () {
    openStreamSuccess = t.step_func(function (fs) {
        assert_true("write" in fs, "write exists");
        fs.close();
        t.done();
    });
    openStreamError = t.step_func(function (error) {
        assert_unreached("openStream() error callback invoked: name: " + error.name + ", msg: " + error.message);
    });

    resolveSuccess = t.step_func(function (dir) {
        file = dir.createFile(fsTestFileName);
        file.openStream("rw", openStreamSuccess, openStreamError);
    });
    resolveError = t.step_func(function (error) {
        assert_unreached("resolve() error callback invoked: name: " + error.name + ", msg: " + error.message);
    });

    prepareForTesting(t,
        function () {
            tizen.filesystem.resolve("documents", resolveSuccess, resolveError, "rw");
        }
    );
});

</script>
</body>
</html>
